Move data processing functions to separate modules
